Move remaining DNF-related code to the DNF payload
Move all methods and properties for working with repositories, groups and RPM
sources to the class DNFPayload.
